Add new product (or increase quantity) to the cart
First, use Product.findByPk(prodId) to fetch the whole product data to
be added to the cart.
Second, use req.uesr.getCart() to get the cart associated with the user.
Third, use cart.hasProduct(product) to check if this product is already
in the cart or not.
Fourth, if the product is already in the cart, use
product.cartItem.quantity to fetch the quantity of the product for
increment.
Fifth, use cart.addProduct(product, { through: {quantity} }) to update
the new quantity.

Lec.164, Adding New Products to the Cart
Lec.165, Adding Existing Products & Retrieving Cart Products