Add logic to update the password
First, add a hidden input for holding password Token.
Second, add postNewPassword controller and set a POST route to it.
Third, find the user by id and verify the token and see if it expires
or not.
Fourth, generate the hash password and save it back to DB.

Lec.277, Adding Logic to Update the Password